Trying some distro-aware type code

+++ CSDBuild.sh
@@ -37,6 +37,30 @@ if [ ${CONFREE} -eq 0 ]; then
 	exit 2
 fi
 
+# Check the pre-reqs
+DISTRO=$(lsb_release -i -s)
+REL=$(lsb_release -r -s)
+if [ ${DISTRO} == "Ubuntu" ]; then
+	# It's an Ubuntu of sorts
+	if [ ${REL} == "16.04" -o ${REL} == "16.10" ]; then
+		PREREQS="openconnect network-manager-openconnect network-manager-openconnect-gnome curl"
+	fi
+fi
+
+# No prereqs and no distforce, die
+if [ -z ${PREREQS} -a ${DISTFORCE} -ne 1 ]; then
+	echo "Unsupported version and DISTFORCE=1 not set... bailing out";
+	exit 1
+fi
+
+# Check each pre-req
+for prereq in ${PREREQS}; do
+	if [ ${DISTRO} == "Ubuntu" ]; then
+		dpkg -l ${prereq} || sudo apt-get install ${prereq}
+	fi
+fi
+
+
 # Download the tarball
 TARBALL=$(mktemp)
 curl -o ${TARBALL} ${PREDEPLOY}
